requirements.txt
crewai
cdp-sdk
openai